part of second stage

+++ app/src/main/java/com/example/hw_first/Models/SkillInfo.kt
@@ -2,5 +2,5 @@ package com.example.hw_first.Models
 
 class SkillInfo(
     val studentSkill: String,
-    val studentExperience: Double
+    val studentExperience: Float
 )
